requirements.txt
# Core
pydantic>=2.7.0
httpx>=0.27.0
networkx>=3.3
numpy>=1.26.0
pyyaml>=6.0.1

# RAG / Retrieval
FlagEmbedding>=1.2.0
faiss-cpu>=1.8.0             # Use faiss-gpu for production
dspy-ai>=2.4.0

# LLM Inference
vllm>=0.5.0
transformers>=4.43.0
torch>=2.3.0

# Knowledge Graph
neo4j>=5.20.0

# Safety
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
openai>=1.35.0               # For Whisper TTS

# Evaluation
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
